requirements.txt
faster-whisper>=0.10.0
ttkbootstrap>=1.10.1
Pillow>=10.4
